requirements.txt
Pillow>=10.0.0
opencv-python>=4.10.0
